Updated Accounts Page and Account Settings Modal
