+++ components/brave_wallet_ui/components/desktop/popup-modals/edit-visible-assets-modal/style.ts
@@ -1,0 +1,44 @@
+import styled from 'styled-components'
+import { LoaderIcon } from 'brave-ui/components/icons'
+
+export const StyledWrapper = styled.div`
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  width: 100%;
+  padding: 0px 15px 15px 15px;
+  min-height: 320px;
+`
+
+export const WatchlistScrollContainer = styled.div`
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: flex-start;
+  width: 100%;
+  overflow-y: scroll;
+  overflow-x: hidden;
+  position: relative;
+  min-height: 225px;
+  max-height: 225px;
+  margin-bottom: 24px;
+  box-sizing: border-box;
+  padding-right: 4px;
+`
+
+export const LoadingWrapper = styled.div`
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  width: 100%;
+  height: 100%;
+`
+
+export const LoadIcon = styled(LoaderIcon)`
+  color: ${p => p.theme.color.interactive08};
+  height: 70px;
+  width: 70px;
+  opacity: .4;
+`
