+++ components/brave_wallet_ui/components/desktop/views/accounts/index.tsx
@@ -3,9 +3,7 @@ import * as React from 'react'
 import {
   WalletAccountType,
   RPCTransactionType,
-  AccountSettingsNavTypes,
-  TokenInfo,
-  AccountAssetOptionType
+  AccountSettingsNavTypes
 } from '../../../../constants/types'
 import { reduceAddress } from '../../../../utils/reduce-address'
 import { copyToClipboard } from '../../../../utils/copy-to-clipboard'
@@ -36,7 +34,6 @@ import {
   WalletInfoLeftSide,
   QRCodeIcon,
   EditIcon,
-  EditButtonRow,
   SubviewSectionTitle
 } from './style'
 
@@ -51,39 +48,29 @@ import {
 } from '../../'
 
 export interface Props {
-  userWatchList: string[]
   accounts: WalletAccountType[]
-  userAssetList: AccountAssetOptionType[]
   transactions: (RPCTransactionType | undefined)[]
-  fullAssetList: TokenInfo[]
   privateKey: string
   onViewPrivateKey: (address: string) => void
   onDoneViewingPrivateKey: () => void
   toggleNav: () => void
   onClickBackup: () => void
   onClickAddAccount: () => void
-  onUpdateVisibleTokens: (list: string[]) => void
   onUpdateAccountName: (name: string) => void
-  fetchFullTokenList: () => void
   onRemoveAccount: (address: string) => void
 }
 
 function Accounts (props: Props) {
   const {
     accounts,
-    userAssetList,
     transactions,
-    userWatchList,
-    fullAssetList,
     privateKey,
     onViewPrivateKey,
     onDoneViewingPrivateKey,
     toggleNav,
     onClickBackup,
     onClickAddAccount,
-    onUpdateVisibleTokens,
     onUpdateAccountName,
-    fetchFullTokenList,
     onRemoveAccount
   } = props
 
@@ -131,18 +118,7 @@ function Accounts (props: Props) {
   }
 
   const onChangeTab = (id: AccountSettingsNavTypes) => {
-    if (id === 'watchlist') {
-      fetchFullTokenList()
-    }
     setEditTab(id)
-  }
-
-  const toggleShowEditWatchlist = () => {
-    if (!showEditModal) {
-      fetchFullTokenList()
-    }
-    setShowEditModal(!showEditModal)
-    setEditTab('watchlist')
   }
 
   const onShowEditModal = () => {
@@ -256,14 +232,6 @@ function Accounts (props: Props) {
               icon={item.asset.icon}
             />
           )}
-          <EditButtonRow>
-            <AddButton
-              buttonType='secondary'
-              onSubmit={toggleShowEditWatchlist}
-              text={locale.accountsEditVisibleAssets}
-              editIcon={true}
-            />
-          </EditButtonRow>
           <SubviewSectionTitle>{locale.transactions}</SubviewSectionTitle>
           <SubDivider />
           {transactions?.map((transaction) =>
@@ -280,23 +248,19 @@ function Accounts (props: Props) {
       )}
       {showEditModal && selectedAccount &&
         <AccountSettingsModal
-          userAssetList={userAssetList}
           title={locale.account}
           account={selectedAccount}
           onClose={onCloseEditModal}
           onUpdateAccountName={onUpdateAccountName}
-          onUpdateVisibleTokens={onUpdateVisibleTokens}
           onCopyToClipboard={onCopyToClipboard}
           onChangeTab={onChangeTab}
+          onToggleNav={toggleNav}
           onRemoveAccount={onRemoveAccount}
           onViewPrivateKey={onViewPrivateKey}
           onDoneViewingPrivateKey={onDoneViewingPrivateKey}
-          onToggleNav={toggleNav}
           privateKey={privateKey}
           tab={editTab}
           hideNav={false}
-          fullAssetList={fullAssetList}
-          userWatchList={userWatchList}
         />
       }
     </StyledWrapper>
