+++ components/brave_wallet_ui/constants/locale.ts
@@ -187,6 +187,9 @@ const locale = {
   accountSettingsRemove: 'Remove Account',
   watchlistButton: 'Save Account Watchlist',
   watchListSearchPlaceholder: 'Search tokens or contract address',
+  accountSettingsDisclaimer: 'Warning: Never disclose this key. Anyone with your private key can steal any assets held in your account.',
+  accountSettingsShowKey: 'Show Key',
+  accountSettingsHideKey: 'Hide Key',
 
   // AmountPresets
   preset25: '25%',
