+++ components/brave_wallet_ui/constants/types.ts
@@ -91,7 +91,6 @@ export type AddAccountNavTypes =
 
 export type AccountSettingsNavTypes =
   | 'details'
-  | 'watchlist'
   | 'privateKey'
 
 export type BuySendSwapTypes =
