+++ components/brave_wallet_ui/options/account-settings-nav-options.ts
@@ -5,10 +5,6 @@ export const AccountSettingsNavOptions: TopTabNavObjectType[] = [
   {
     id: 'details',
     name: locale.accountSettingsDetails
-  },
-  {
-    id: 'watchlist',
-    name: locale.accountSettingsWatchlist
   }
 ]
